# Conflicts:
#	build/libjpeg/libjpeg.vcxproj
#	build/libpng/libpng.vcxproj
#	build/miniblink.sln
#	build/miniblink/miniblink.vcxproj
#	cc/layers/CompositingLayer.cpp
#	cc/raster/RasterTaskWorkerThreadPool.cpp
#	cc/tiles/CompositingTile.cpp
#	gen/blink/bindings/core/v8/V8MediaQueryList.cpp
#	net/WebURLLoaderInternal.h
#	net/WebURLLoaderManager.cpp
#	net/WebURLLoaderManager.h

+++ content/browser/NavigationController.cpp
@@ -36,7 +36,7 @@ int NavigationController::historyForwardListCount()
 void NavigationController::navigate(int offset)
 {
     int pos = m_currentOffset + offset;
-    if (pos < 0 && pos > m_items.size() - 1)
+    if (pos < 0 && pos > (int)(m_items.size() - 1))
         return;
     blink::WebHistoryItem item = m_items[pos];
 #ifdef DEBUG
@@ -66,7 +66,7 @@ void NavigationController::insertOrReplaceEntry(const blink::WebHistoryItem& ite
     switch (type) {
     case blink::WebStandardCommit:
         ++m_currentOffset;
-        ASSERT(0 <= m_currentOffset && m_currentOffset <= m_items.size());
+        ASSERT(0 <= m_currentOffset && m_currentOffset <= (int)m_items.size());
         if (m_currentOffset == m_items.size()) {
             m_items.append(item);
         } else {
